MINOR: Renames IBCT files
