+++ @here/harp-test-utils/index.ts
@@ -6,5 +6,3 @@
 
 export * from "./lib/TestDataUtils";
 export * from "./lib/TestUtils";
-export * from "./lib/rendering/FeedbackCli";
-export * from "./lib/rendering/FeedbackServer";
