+++ @here/harp-test-utils/index.web.ts
@@ -6,4 +6,4 @@
 
 export * from "./lib/TestDataUtils.web";
 export * from "./lib/TestUtils";
-export * from "./lib/rendering/ImageBasedTestHelper";
+export * from "./lib/rendering/RenderingTestHelper";
