+++ @here/harp-test-utils/lib/rendering/ReferenceImageLocator.ts
