+++ @here/harp-test-utils/lib/rendering/RenderingTestResultCommon.ts
